Segregate the routers into two types
